Add a testcase to test that a user can view their own posts
An array with all the user's posts is returned - this implementation was chosen over echoeing
the string because we do not know at the moment how exactly the frontend will be implemented.
It also seems that we might want to pass the post string into some function to prettify it with
CSS and HTML so a simple return seems more appropriate.